2.1.2 Создаст модуль для утилит, вынесет в него общие функции.
